requirements.txt
Django==2.1.5
jiphy==1.2.2
PyExecJS==1.5.1
pytz==2018.9
six==1.12.0
